conda_requirements.txt
torch
pandas
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bzip2=1.0.6=h9a117a8_4
ca-certificates=2017.08.26=h1d4fec5_0
cairo=1.14.12=h77bcde2_0
certifi=2018.1.18=py35_0
cffi=1.11.4=py35h9745a5d_0
cycler=0.10.0=py35hc4d5149_0
dbus=1.12.2=hc3f9b76_1
decorator=4.2.1=py35_0
expat=2.2.5=he0dffb1_0
ffmpeg=3.4=h7264315_0
fontconfig=2.12.4=h88586e7_1
freetype=2.8=hab7d2ae_1
glib=2.53.6=h5d9569c_2
graphite2=1.3.10=hf63cedd_1
gst-plugins-base=1.12.4=h33fb286_0
gstreamer=1.12.4=hb53b477_0
harfbuzz=1.7.4=hc5b324e_0
hdf5=1.10.1=h9caa474_1
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.0=hc7b2577_8
ipython=6.2.1=py35hd850d2a_1
ipython_genutils=0.2.0=py35hc9e07d0_0
jasper=1.900.1=hd497a04_4
jedi=0.11.1=py35_0
jpeg=9b=h024ee3a_2
libedit=3.1=heed3624_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran-ng=7.2.0=h9f7466a_2
libopus=1.2.1=hb9ed12e_0
libpng=1.6.34=hb9fc6fc_0
libprotobuf=3.4.1=h5b8497f_0
libstdcxx-ng=7.2.0=h7a57d05_2
libtiff=4.0.9=h28f6b97_0
libvpx=1.6.1=h888fd40_0
libxcb=1.12=hcd93eb1_4
libxml2=2.9.7=h26e45fe_0
matplotlib=2.1.2=py35h0e671d2_0
mkl=2018.0.1=h19d6760_4
ncurses=6.0=h9df7e31_2
numpy=1.14.1=py35h3dfced4_1
olefile=0.45.1=py35_0
opencv=3.3.1=py35h6cbbc71_1
openssl=1.0.2n=hb7f436b_0
pandas=0.22.0=py35hf484d3e_0
parso=0.1.1=py35h1b200a3_0
patsy=0.5.0=py35_0
pcre=8.41=hc27e229_1
pexpect=4.4.0=py35_0
pickleshare=0.7.4=py35hd57304d_0
pillow=5.0.0=py35h3deb7b8_0
pip=9.0.1=py35h7e7da9d_4
pixman=0.34.0=hceecf20_3
prompt_toolkit=1.0.15=py35hc09de7a_0
ptyprocess=0.5.2=py35h38ce0a3_0
pycparser=2.18=py35h61b3040_1
pygments=2.2.0=py35h0f41973_0
pyparsing=2.2.0=py35h041ed72_1
pyqt=5.6.0=py35h0e41ada_5
python=3.5.4=hc3d631a_27
python-dateutil=2.6.1=py35h90d5b31_1
pytz=2018.3=py35_0
qt=5.6.2=hd25b39d_14
readline=7.0=ha6073c6_4
scikit-learn=0.19.1=py35hbf1f462_0
scipy=1.0.0=py35hcbbe4a2_0
seaborn=0.8.1=py35h04cba02_0
setuptools=38.5.1=py35_0
simplegeneric=0.8.1=py35_2
sip=4.18.1=py35h9eaea60_2
six=1.11.0=py35h423b573_1
sqlite=3.22.0=h1bed415_0
statsmodels=0.8.0=py35haa9d50b_0
tk=8.6.7=hc745277_3
tornado=4.5.3=py35_0
traitlets=4.3.2=py35ha522a97_0
wcwidth=0.1.7=py35hcd08066_0
wheel=0.30.0=py35hd3883cf_1
xz=5.2.3=h55aa19d_2
zlib=1.2.11=ha838bed_2
